adding recent code

+++ base/views.py
@@ -1,9 +1,12 @@
 from django.shortcuts import render,redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate,login,logout
-from .models import Room,Topic
+from django.contrib.auth.forms import UserCreationForm
+from .models import Room,Topic,Message
 from .forms import RoomForm
 
 # rooms=[
@@ -15,8 +18,11 @@ from .forms import RoomForm
 # Create your views here.
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
     if request.method=="POST":
-        username=request.POST.get('username')
+        username=request.POST.get('username').lower()
         password=request.POST.get('password')
         try:
             user = User.objects.get(username=username) 
@@ -30,9 +36,29 @@ def loginPage(request):
         else:
             message.error(request,"Username or password does not exist")
         
-    c={}
+    c={'page':page}
     return render(request,'base/login_register.html',c)
 
+@login_required(login_url='login')
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+    c = {"form":form}
+    if request.method=="POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,"An error occurred during registration")
+    return render(request,'base/login_register.html',c)
+    
 def home(request):
     q = request.GET.get('q')  if request.GET.get('q') !=None else ''
     
@@ -41,18 +67,27 @@ def home(request):
         )
     topics = Topic.objects.all()
     room_count = rooms.count()
-    
-    c = {'rooms':rooms,"topics":topics,"room_count":room_count} 
+    roomMessages = Message.objects.filter(Q(room__topic__name__icontains=q))
+    print(roomMessages)
+    c = {'rooms':rooms,"topics":topics,"room_count":room_count,'room_messages':roomMessages} 
     return render(request,'base/home.html',c)
 
 def room(request,pk):
     room = Room.objects.get(id=pk)
-    # for i in rooms:
-    #     if i['id']== int(pk):
-    #         room=i    
-    c = {'room':room}
+    roomMessages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
+    if request.method=='POST':
+        message = Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body')
+        )
+        room.participants.add(request.user )
+        return redirect('room',pk=room.id)
+    c = {'room':room,'roomMessages':roomMessages,'participants':participants}
     return render(request,"base/room.html",c)
 
+@login_required(login_url='login')
 def createRoom(request):
     if request.method=="POST":
         form = RoomForm(request.POST)
@@ -64,8 +99,11 @@ def createRoom(request):
         c={'form':form}
         return render(request,'base/room_form.html',c)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
+    if request.user!=room.host:
+        return HttpResponse('Your are not allowed Here')
     if request.method=="POST":
         form = RoomForm(request.POST,instance=room)
         if form.is_valid():
@@ -75,11 +113,25 @@ def updateRoom(request,pk):
     c = {'form':form}
     return render(request,'base/room_form.html',c)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+    if request.user!=room.host:
+        return HttpResponse('Your are not allowed Here')
     if request.method=='POST':
         room.delete()
         return redirect('home')
     c={'obj':room}
     return render(request,'base/delete.html',c)
 
+@login_required(login_url='login')
+def deleteMessage(request,pk):
+    message = Message.objects.get(id=pk)
+    if request.user!=message.user:
+        return HttpResponse('Your are not allowed Here')
+    if request.method=='POST':
+        message.delete()
+        return redirect('home')
+    c={'obj':message}
+    return render(request,'base/delete.html',c)
+
